+++ NaLaPla/Plan.cs
@@ -1,0 +1,10 @@
+namespace NaLaPla
+{
+    public class Plan {
+        public string? description;
+        public int planLevel;        
+        public List<Plan>? planSteps;
+
+        public Plan? parent;
+    }
+}
