+++ NaLaPla/Program.cs
@@ -30,15 +30,7 @@
                 };
 
             await ExpandPlan(basePlan);
-            WritePlan(basePlan);
-        }
-
-        static void WritePlan(Plan plan) {
-            var description = plan.description.PadLeft(plan.description.Length + (5*plan.planLevel));
-            Console.WriteLine(description);
-            foreach (var subPlan in plan.planSteps) {
-                WritePlan(subPlan);
-            }
+            Util.WritePlan(basePlan);
         }
 
         static async Task ExpandPlan(Plan planToExpand) {
@@ -72,7 +64,7 @@
                 temperature: 0.6);
 
             var rawPlan = result.ToString();
-            var subTasks = ParseList(rawPlan);
+            var subTasks = Util.ParseList(rawPlan);
             return subTasks;
         }
 
@@ -86,55 +78,5 @@
             Console.WriteLine(firstPrompt);
             return firstPrompt;
         }
-
-        
-        static List<string> ParseList(string itemString) {
-
-            // Assume list is like: "1. this 2. that 3. other"
-            var list = itemString.Split("\n").ToList();
-
-            list = list.Select((n) => {
-                var breakPos = n.IndexOf(". ");
-                if (breakPos > -1) {
-                    return n.Substring(breakPos+2);
-                }
-                return n;
-            }).Distinct().ToList();
-            
-            list.RemoveAll(s => string.IsNullOrEmpty(s));
-
-            /*
-            // Remove the first number
-            list.RemoveAt(0);
-
-            char[] ch = {'.', '\n'};
-            list = list.Select((n) => {
-                var breakPos = n.LastIndexOfAny(ch);
-                if (breakPos > -1) {
-                    return n.Substring(0, breakPos);
-                }
-                return n;
-            }).Distinct().ToList();
-            */
-            return list;
-        }
-
-        static List<string> Test() {
-            var parse1 = "1. Choose a name for the company.\n\n2. Create a logo for the company.\n\n3. Develop a business plan.\n\n4. Raise capital.\n\n5. Find a manufacturing partner.\n\n6. Build a prototype.\n\n7. Test the prototype.\n\n8. Launch the production version of the car.\n\n9. Sell the car.\n\n";
-            var parse2 = "\n\n1. Research the electric car market. This includes understanding the current landscape of electric car manufacturers, what consumers want in an electric car, and what features are most important to them.\n\n2. Develop a business plan for your electric car company. This should include your company’s mission and vision, as well as your target market and marketing strategy.\n\n3. Raise capital for your electric car company. This may involve seeking out investors or applying for loans.\n\n4. Hire a team of experts to help you build your electric car company. This may include engineers, designers, and marketing professionals.\n\n5. Develop a prototype of your electric car. This should be a working model that can be used to test and refine your design.\n\n6. Launch your electric car company. This includes marketing your car to consumers and setting up a sales and distribution network.";
-            var parse3 = "\n\n1. Hire staff\n2. Set up the production line";
-            var list = ParseList(parse1);
-            list = ParseList(parse2);
-            list = ParseList(parse3);
-            return list;
-        }
-    }
-
-    class Plan {
-        public string? description;
-        public int planLevel;        
-        public List<Plan>? planSteps;
-
-        public Plan? parent;
     }
   }
